7.5 Разделит редьюсер по зонам ответственности: состояние приложения (filters), загружаемые данные (offers)
