+++ src/reducer/offers/offers.js
@@ -1,45 +1,24 @@
-import {extend} from "./utils.js";
-import {SortTypes} from "./const.js";
+import {extend} from "../../utils.js";
 import {getCities} from "./selectors.js";
-import {createOffer, createReview} from "./adapters/adapters.js";
+import {createOffer, createReview} from "../../adapters/adapters.js";
+import {ActionCreator as FiltersActionCreator} from "../filters/filters.js";
 
 const initialState = {
-  city: null,
   offers: [],
-  sortType: SortTypes.POPULAR,
-  activeCard: null,
   currentOfferReviews: [],
   currentOffersNearby: [],
 };
 
 const ActionType = {
-  CHANGE_CITY: `CHANGE_CITY`,
   SET_ALL_OFFERS: `SET_ALL_OFFERS`,
-  CHANGE_SORT_TYPE: `CHANGE_SORT_TYPE`,
-  CHANGE_ACTIVE_CARD: `CHANGE_ACTIVE_CARD`,
   CHANGE_CURRENT_OFFER_REVIEWS: `CHANGE_CURRENT_OFFER_REVIEWS`,
   CHANGE_CURRENT_OFFERS_NEARBY: `CHANGE_CURRENT_OFFERS_NEARBY`,
 };
 
 const ActionCreator = {
-  changeCity: (city) => ({
-    type: ActionType.CHANGE_CITY,
-    payload: city,
-  }),
-
   setAllOffers: (offers) => ({
     type: ActionType.SET_ALL_OFFERS,
     payload: offers,
-  }),
-
-  changeSortType: (sortType) => ({
-    type: ActionType.CHANGE_SORT_TYPE,
-    payload: sortType,
-  }),
-
-  changeActiveCard: (offer) => ({
-    type: ActionType.CHANGE_ACTIVE_CARD,
-    payload: offer,
   }),
 
   changeCurrentOfferReviews: (reviews) => ({
@@ -67,7 +46,7 @@ const Operation = {
         const cities = getCities(getState());
         if (cities.length) {
           const city = cities[0];
-          dispatch(ActionCreator.changeCity(city));
+          dispatch(FiltersActionCreator.changeCity(city));
         }
       });
   },
@@ -99,24 +78,9 @@ const Operation = {
 
 const reducer = (state = initialState, action) => {
   switch (action.type) {
-    case ActionType.CHANGE_CITY:
-      return extend(state, {
-        city: action.payload,
-      });
-
     case ActionType.SET_ALL_OFFERS:
       return extend(state, {
         offers: action.payload,
-      });
-
-    case ActionType.CHANGE_SORT_TYPE:
-      return extend(state, {
-        sortType: action.payload,
-      });
-
-    case ActionType.CHANGE_ACTIVE_CARD:
-      return extend(state, {
-        activeCard: action.payload,
       });
 
     case ActionType.CHANGE_CURRENT_OFFER_REVIEWS:
